+++ packages/cubejs-api-gateway/src/gateway.ts
@@ -502,30 +502,23 @@ export class ApiGateway {
           compilerApi,
           query
         );
-
-      const preAggregationVersionEntries = preAggregationPartitions &&
+        
+      const versionEntriesResult = preAggregationPartitions &&
         await orchestratorApi.getPreAggregationVersionEntries(
           context,
           preAggregationPartitions,
           compilerApi.preAggregationsSchema
         );
 
-      const mergePartitionsAndVersionEntries = () => {
-        const preAggregationVersionEntriesByName = preAggregationVersionEntries.reduce((obj, versionEntry) => {
-          if (!obj[versionEntry.table_name]) obj[versionEntry.table_name] = [];
-          obj[versionEntry.table_name].push(versionEntry);
-          return obj;
-        }, {});
-
-        return ({ preAggregation, partitions, ...props }) => ({
-          ...props,
-          preAggregation,
-          partitions: partitions.map(partition => {
-            partition.versionEntries = preAggregationVersionEntriesByName[partition.sql?.tableName] || [];
-            return partition;
-          }),
-        });
-      };
+      const mergePartitionsAndVersionEntries = () => ({ preAggregation, partitions, ...props }) => ({
+        ...props,
+        preAggregation,
+        partitions: partitions.map(partition => {
+          partition.versionEntries = versionEntriesResult?.versionEntriesByTableName[partition.sql?.tableName] || [];
+          partition.structureVersion = versionEntriesResult?.structureVersionsByTableName[partition.sql?.tableName];
+          return partition;
+        }),
+      });
 
       res({
         preAggregationPartitions: preAggregationPartitions.map(mergePartitionsAndVersionEntries())