+++ packages/cubejs-playground/src/playground/index.ts
@@ -1,8 +1,6 @@
 import './index.less';
 
 export * from './components/QueryBuilder';
-export * from '../components/SecurityContext/SecurityContextProvider';
-export * from '../components/LivePreviewContext/LivePreviewContextProvider';
 export * from '../hooks';
 export { fetchPoll } from '../utils';
 export { setTracker, setTelemetry } from '../events';