+++ packages/cubejs-playground/src/shared/helpers.ts
@@ -86,12 +86,12 @@ type RequestOptions = {
   token?: string;
   body?: Record<string, any>;
   headers?: Record<string, string>;
-}
+};
 
 export async function request(
   endpoint: string,
   method: string = 'GET',
-  options: RequestOptions
+  options: RequestOptions = {}
 ) {
   const { body, token } = options;
 
@@ -112,6 +112,56 @@ export async function request(
 
   return {
     ok: response.ok,
-    json: await response.json()
+    json: await response.json(),
   };
+}
+
+type OpenWindowOptions = {
+  url: string;
+  width?: number;
+  height?: number;
+  title?: string;
+};
+
+export function openWindow({
+  url,
+  title = '',
+  width = 640,
+  height = 720,
+}: OpenWindowOptions) {
+  const dualScreenLeft =
+    window.screenLeft !== undefined ? window.screenLeft : window.screenX;
+  const dualScreenTop =
+    window.screenTop !== undefined ? window.screenTop : window.screenY;
+
+  const w = window.innerWidth
+    ? window.innerWidth
+    : document.documentElement.clientWidth
+    ? document.documentElement.clientWidth
+    : screen.width;
+  const h = window.innerHeight
+    ? window.innerHeight
+    : document.documentElement.clientHeight
+    ? document.documentElement.clientHeight
+    : screen.height;
+
+  const systemZoom = w / window.screen.availWidth;
+  const left = (w - width) / 2 / systemZoom + dualScreenLeft;
+  const top = (h - height) / 2 / systemZoom + dualScreenTop;
+
+  const newWindow = window.open(
+    url,
+    title,
+    `
+      scrollbars=yes,
+      width=${width / systemZoom}, 
+      height=${height / systemZoom}, 
+      top=${top}, 
+      left=${left}
+    `
+  );
+
+  newWindow?.focus?.();
+
+  return newWindow;
 }