+++ packages/cubejs-query-orchestrator/src/orchestrator/PreAggregations.ts
@@ -94,6 +94,7 @@ type IndexDescription = {
 
 type PreAggregationDescription = {
   preAggregationId: string;
+  priority: number;
   previewSql: QueryWithParams;
   timezone: string;
   indexesSql: IndexDescription[];
@@ -135,7 +136,8 @@ const tablesToVersionEntries = (schema, tables: TableCacheEntry[]): VersionEntry
 
 type PreAggregationLoadCacheOptions = {
   requestId?: string,
-  dataSource: string
+  dataSource: string,
+  tablePrefixes?: string[],
 };
 
 type VersionEntriesObj = {
@@ -174,6 +176,8 @@ class PreAggregationLoadCache {
 
   private dataSource: string;
 
+  private tablePrefixes: string[] | null;
+
   public constructor(
     redisPrefix,
     clientFactory: DriverFactory,
@@ -190,6 +194,7 @@ class PreAggregationLoadCache {
     this.cacheDriver = preAggregations.cacheDriver;
     this.externalDriverFactory = preAggregations.externalDriverFactory;
     this.requestId = options.requestId;
+    this.tablePrefixes = options.tablePrefixes;
     this.versionEntries = {};
     this.tables = {};
   }
@@ -228,6 +233,9 @@ class PreAggregationLoadCache {
     const client = preAggregation.external ?
       await this.externalDriverFactory() :
       await this.driverFactory();
+    if (this.tablePrefixes && client.getPrefixTablesQuery && this.preAggregations.options.skipExternalCacheAndQueue) {
+      return client.getPrefixTablesQuery(preAggregation.preAggregationsSchema, this.tablePrefixes);
+    }
     return client.getTablesQuery(preAggregation.preAggregationsSchema);
   }
 
@@ -286,6 +294,9 @@ class PreAggregationLoadCache {
   }
 
   public async getVersionEntries(preAggregation): Promise<VersionEntriesObj> {
+    if (this.tablePrefixes && !this.tablePrefixes.find(p => preAggregation.tableName.split('.')[1].startsWith(p))) {
+      throw new Error(`Load cache tries to load table ${preAggregation.tableName} outside of tablePrefixes filter: ${this.tablePrefixes.join(', ')}`);
+    }
     const redisKey = this.tablesRedisKey(preAggregation);
     if (!this.versionEntries[redisKey]) {
       this.versionEntries[redisKey] = this.calculateVersionEntries(preAggregation).catch(e => {
@@ -520,7 +531,7 @@ export class PreAggregationLoader {
         await this.loadCache.getVersionEntries(this.preAggregation)
       );
       if (!lastVersion) {
-        throw new Error(`Pre-aggregation table is not found for ${this.preAggregation.tableName} after it was successfully created. It usually means database silently truncates table names due to max name length.`);
+        throw new Error(`Pre-aggregation table is not found for ${this.preAggregation.tableName} after it was successfully created`);
       }
       return this.targetTableName(lastVersion);
     };
@@ -983,6 +994,12 @@ export class PreAggregationLoader {
 }
 
 export class PreAggregationPartitionRangeLoader {
+  protected waitForRenew: boolean;
+
+  protected requestId: string;
+
+  protected dataSource: string;
+
   public constructor(
     private readonly redisPrefix: string,
     private readonly driverFactory: DriverFactory,
@@ -995,10 +1012,50 @@ export class PreAggregationPartitionRangeLoader {
     private readonly loadCache: any,
     private readonly options: any = {}
   ) {
+    this.waitForRenew = options.waitForRenew;
+    this.requestId = options.requestId;
+    this.dataSource = options.dataSource;
   }
 
-  private async loadRangeQuery(rangeQuery) {
-    return this.loadCache.keyQueryResult(rangeQuery, false, 10, 60 * 60); // TODO 60 * 60
+  private async loadRangeQuery(rangeQuery: QueryTuple, partitionRange?: QueryDateRange) {
+    const [query, values, queryOptions]: QueryTuple = rangeQuery;
+
+    return this.queryCache.cacheQueryResult(
+      query,
+      values,
+      [query, values],
+      24 * 60 * 60,
+      {
+        renewalThreshold: this.queryCache.options.refreshKeyRenewalThreshold
+          || queryOptions?.renewalThreshold || 24 * 60 * 60,
+        waitForRenew: this.waitForRenew,
+        priority: this.priority(10),
+        requestId: this.requestId,
+        dataSource: this.dataSource,
+        useInMemory: true,
+        external: queryOptions?.external,
+        renewalKey: partitionRange ? await this.getInvalidationKeyValues(partitionRange) : null
+      }
+    );
+  }
+
+  protected getInvalidationKeyValues(range) {
+    const partitionTableName = PreAggregationPartitionRangeLoader.partitionTableName(
+      this.preAggregation.tableName, this.preAggregation.partitionGranularity, range
+    );
+    return Promise.all(
+      (this.preAggregation.invalidateKeyQueries || []).map(
+        (sqlQuery) => (
+          this.loadCache.keyQueryResult(
+            this.replacePartitionSqlAndParams(sqlQuery, range, partitionTableName), this.waitForRenew, this.priority(10)
+          )
+        )
+      )
+    );
+  }
+
+  protected priority(defaultValue) {
+    return this.preAggregation.priority != null ? this.preAggregation.priority : defaultValue;
   }
 
   private replacePartitionSqlAndParams(
@@ -1107,38 +1164,53 @@ export class PreAggregationPartitionRangeLoader {
 
   private async partitionRanges() {
     const { preAggregationStartEndQueries } = this.preAggregation;
-    let dateRange = this.preAggregation.matchedTimeDimensionDateRange;
-    if (!this.options.skipLoadRangeQuery || !dateRange) {
-      const [startDate, endDate] = await Promise.all(
-        preAggregationStartEndQueries.map(
-          async rangeQuery => PreAggregationPartitionRangeLoader.extractDate(await this.loadRangeQuery(rangeQuery)),
-        ),
-      );
-
-      if (startDate && endDate) {
-        this.logger('PreAggregation Refresh Range', {
-          requestId: this.options.requestId,
-          preAggregationId: this.preAggregation.preAggregationId,
-          refreshRange: [startDate, endDate]
-        });
-      }
-
-      dateRange = PreAggregationPartitionRangeLoader.intersectDateRanges(
-        [startDate, endDate],
-        this.preAggregation.matchedTimeDimensionDateRange,
-      );
-
-      if (!dateRange) {
-        // If there's no date range intersection between query data range and pre-aggregation build range
-        // use last partition so outer query can receive expected table structure.
-        dateRange = [endDate, endDate];
-      }
+    const [startDate, endDate] = await Promise.all(
+      preAggregationStartEndQueries.map(
+        async rangeQuery => PreAggregationPartitionRangeLoader.extractDate(await this.loadRangeQuery(rangeQuery)),
+      ),
+    );
+    let dateRange = PreAggregationPartitionRangeLoader.intersectDateRanges(
+      [startDate, endDate],
+      this.preAggregation.matchedTimeDimensionDateRange,
+    );
+    if (!dateRange) {
+      // If there's no date range intersection between query data range and pre-aggregation build range
+      // use last partition so outer query can receive expected table structure.
+      dateRange = [endDate, endDate];
     }
-
-    return PreAggregationPartitionRangeLoader.timeSeries(
+    let timeSeriesRanges = PreAggregationPartitionRangeLoader.timeSeries(
       this.preAggregation.partitionGranularity,
       dateRange,
     );
+    const wholeSeriesRanges = PreAggregationPartitionRangeLoader.timeSeries(
+      this.preAggregation.partitionGranularity,
+      [startDate, endDate],
+    );
+    const [preciseStartDate, preciseEndDate] = await Promise.all(
+      preAggregationStartEndQueries.map(
+        async (rangeQuery, i) => PreAggregationPartitionRangeLoader.extractDate(
+          await this.loadRangeQuery(
+            rangeQuery, i === 0 ? wholeSeriesRanges[0] : wholeSeriesRanges[wholeSeriesRanges.length - 1]
+          )
+        ),
+      ),
+    );
+    if (preciseStartDate !== startDate || preciseEndDate !== endDate) {
+      dateRange = PreAggregationPartitionRangeLoader.intersectDateRanges(
+        [preciseStartDate, preciseEndDate],
+        this.preAggregation.matchedTimeDimensionDateRange,
+      );
+      if (!dateRange) {
+        // If there's no date range intersection between query data range and pre-aggregation build range
+        // use last partition so outer query can receive expected table structure.
+        dateRange = [preciseEndDate, preciseEndDate];
+      }
+      timeSeriesRanges = PreAggregationPartitionRangeLoader.timeSeries(
+        this.preAggregation.partitionGranularity,
+        dateRange,
+      );
+    }
+    return timeSeriesRanges;
   }
 
   private static checkDataRangeType(range: QueryDateRange) {
@@ -1274,7 +1346,13 @@ export class PreAggregations {
         loadCacheByDataSource[dataSource] =
           new PreAggregationLoadCache(this.redisPrefix, () => this.driverFactory(dataSource), this.queryCache, this, {
             requestId: queryBody.requestId,
-            dataSource
+            dataSource,
+            tablePrefixes:
+              // Can't reuse tablePrefixes for shared refresh scheduler cache
+              !queryBody.preAggregationsLoadCacheByDataSource ?
+                preAggregations
+                  .filter(p => (p.dataSource || 'default') === dataSource)
+                  .map(p => p.tableName.split('.')[1]) : null
           });
       }