+++ packages/cubejs-query-orchestrator/src/orchestrator/QueryQueue.js
@@ -209,7 +209,15 @@ export class QueryQueue {
     const redisClient = await this.queueDriver.createConnection();
     try {
       const query = await redisClient.cancelQuery(queryKey);
-      if (query) await this.sendCancelMessageFn(query);
+
+      if (query) {
+        this.logger('Cancelling query manual', {
+          queryKey: query.queryKey,
+          queuePrefix: this.redisQueuePrefix,
+          requestId: query.requestId
+        });
+        await this.sendCancelMessageFn(query);
+      }
 
       return true;
     } finally {